pose-detection use local movenet model file
the argument "modelConfig.modelUrl" is only supported for string but iohandler
when I try to  resolve the problem of area/countries that don't have access to the model hosted on tf.hub. and I passed the modelUrl like a iohandler，but cause an error indexOf is not undefined. so i put the url check code to resolve this problem